requirements.txt
streamlit==1.32.2
streamlit-option-menu==0.3.12
google-ai-generativelanguage==0.4.0
google-api-core==2.18.0
google-auth==2.29.0
google-generativeai==0.4.1
googleapis-common-protos==1.63.0